Fix segfault when calling yaconv more than once
yaconv_init_once relies on checking if filter_buf is null to detect if it is
allocated. The problem is that calling free on filter_buf does not set said
pointer to null. This commit fixes this by setting said pointer to null upon deinit.

This commit also frees auxinfo to avoid a memory leak.